moved cgi and mime parser to new files
